test: useProducts, useCartItems 훅 RTL 테스트 구현
CartItemCount.test.tsx -> useCartItems.test.tsx
ProductListFilterAndSort.test.tsx -> useProducts.test.tsx

App 컴포넌트 기반으로 구현된 RTL 테스트를 커스텀 훅 기반의 RTL 테스트로 변경
-> 훨씬 테스트하기 편해진듯, 가독성이 좋아짐, 그리고 data-testid 같은 걸 굳이 요소에 추가해줄 필요가 없어짐
